refactor(#155): introduce dir for IDB schemes
