+++ src/requests/backend/translations/idb/schema/index.ts
@@ -1,0 +1,4 @@
+import v1 from './v1';
+import v2 from './v2';
+
+export const IDBTranslationSchemes = [v1, v2];
